Refractor RISC-V emit_branch helpers
Add support for conditional branches by redefining mambo_cond for
RISC-V. On RISC-V a conditional branch instruction takes two registers for
comparison, so they need to be passed for emmiting a conditional branch
instruction.
For RISC-V targets mambo_cond is a struct containing two registers and a
condition code. The riscv_cond_code enum is similar to the branch_condition
enum exept for the additional AL condition which makes it compatible if
checked for AL before.